Forwarding logic
This was a little tricky since I didn't know why it wasn't working when I was taking a lot of the logic from the slides.

1 ahead:
1 ahead checks to see if the reg write from EX/MEM is enabled and also checking to see if we're writing to $0. After doing some digging, I learned that 10 and 01 represent that the data are 1 and 2 cycles ahead respectively so I've commented out the code, but kept it in there just for the sake of knowing that it came from the slides.

2 ahead:
similar logic to the previous note, except that we check to see reg write in MEM/WB is enabled.

Testbench:
Manually did some operations to see what the outputs would be. added some verification logic for the sake of making life easier. We should be all set, but the only thing stopping us is my own mistakes in calculations lol. should be all set.

pushing to main cuz im feeling confident (i am deeply scared for my life right now.)